+++ src/wboxtest/block/ramdisk.c
@@ -29,8 +29,8 @@ static void * ramdisk_setup(struct wboxtest_t * wbt)
 
 	length = sprintf(json,
 		"{\"blk-ramdisk@999\":{\"address\":%lld,\"size\":%lld}}",
-		(virtual_addr_t)(pdat->rambuf),
-		(virtual_size_t)(SZ_1M));
+		(unsigned long long)((virtual_addr_t)pdat->rambuf),
+		(unsigned long long)((virtual_size_t)SZ_1M));
 	probe_device(json, length, NULL);
 
 	pdat->blk = search_block("blk-ramdisk.999");
